#210 finished - added doctor home page, renamed clinic admin home page, removed some unused imports
