+++ frontend/src/components/DoctorComponents/PatientList.js
@@ -1,8 +1,5 @@
 import React from 'react'
 import TableCell from '@material-ui/core/TableCell';
-
-import TableHead from '@material-ui/core/TableHead';
-import TablePagination from '@material-ui/core/TablePagination';
 import TableRow from '@material-ui/core/TableRow';
