+++ frontend/src/components/DoctorComponents/PatientMedicalRecord.js
@@ -2,8 +2,6 @@ import React from 'react'
 import ExpansionPanelDetails from '@material-ui/core/ExpansionPanelDetails';
 import Typography from '@material-ui/core/Typography';
 import Grid from '@material-ui/core/Grid';
-import Paper from '@material-ui/core/Paper';
-import Container from '@material-ui/core/Container';
 
 
 export default function PatientMedicalRecord(props) {
